Fixed whitespace according to developers
Performance is still better with all the fixes and stuff, but the
JavaScript now takes a bit longer to load than with the optimizations.
However, we now have more comments and stuff so maybe more developers
will know what's going on when they look at the code and be more willing
to help? I don't know. Also, maybe optimize the JavaScript and HTML/CSS
on the actual site, but not in the GitHub, that way everyone is happy?
This would mean we could start testing more browsers.